requirements.txt
loguru==0.6.0
Pillow==9.5.0
PyYAML==6.0
torch==2.0.0
torchvision==0.15.0
tqdm==4.65.0
